htb/challenges/web/Saturn/web_saturn/challenge/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
SafeURL-Python==1.3
Werkzeug==3.0.1
